get_stream/sample_chat/scripts/token_generator/pyproject.toml
[project]
name = "token-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.2.1",
    "stream-chat>=4.28.0",
]
